+++ systore_purchase_supply_dashboard/models/__init__.py
@@ -2,4 +2,5 @@ from . import supplier_credit
 from . import purchase_payment
 from . import stock_warehouse
 from . import supply_snapshot
+from . import supply_debt_report
 from . import supply_dashboard